requirements.txt
discord
openai
python-dotenv
Flask
